+++ src/api/controllers/GetUser.ts
@@ -1,0 +1,27 @@
+import { Request } from 'express';
+
+import { StatusCode, JSONResponse, ErrorMessage } from '../../types/types';
+import BaseController from '../BaseController';
+
+class GetUser extends BaseController {
+  async handleRequest(req: Request): Promise<{ statusCode: StatusCode; jsonResponse: JSONResponse }> {
+    const { userId } = req.params;
+
+    const userIdInt = parseInt(userId);
+    if (isNaN(userIdInt)) {
+      return this.format(StatusCode.badRequest, ErrorMessage.INVALID_USER_ID);
+    }
+
+    if (req.user?.getId() !== userIdInt) {
+      return this.format(StatusCode.forbidden, ErrorMessage.UNAUTHORIZED);
+    }
+
+    try {
+      return this.format(StatusCode.ok, null, req.user.truncate());
+    } catch (error) {
+      return this.format(StatusCode.internalServerError, ErrorMessage.ERROR_FINDING_USER);
+    }
+  }
+}
+
+export default GetUser;
