+++ src/api/controllers/RemoveConversationMember.ts
@@ -1,0 +1,40 @@
+import { Request } from 'express';
+
+import { StatusCode, JSONResponse, ErrorMessage } from '../../types/types';
+import Conversation from '../../models/Conversation';
+import BaseController from '../BaseController';
+
+class RemoveConversationMember extends BaseController {
+  async handleRequest(req: Request): Promise<{ statusCode: StatusCode; jsonResponse: JSONResponse }> {
+    const userId = req.user?.getId();
+    const { conversationId } = req.params;
+
+    if (!userId) {
+      return this.format(StatusCode.badRequest, ErrorMessage.UNAUTHENTICATED);
+    }
+
+    const conversationIdInt = parseInt(conversationId, 10);
+    if (isNaN(conversationIdInt)) {
+      return this.format(StatusCode.badRequest, ErrorMessage.INVALID_CONVO_ID);
+    }
+
+    try {
+      const conversation = await Conversation.findById(conversationIdInt);
+      if (!conversation) {
+        return this.format(StatusCode.badRequest, ErrorMessage.CONVO_DOES_NOT_EXIST);
+      }
+
+      const isUserInConversation = await conversation.checkIfHasUser(userId);
+      if (!isUserInConversation) {
+        return this.format(StatusCode.forbidden, ErrorMessage.USER_NOT_IN_CONVO);
+      }
+
+      await conversation.removeUser(userId);
+      return this.format(StatusCode.ok, null, 'User removed from conversation');
+    } catch (error) {
+      return this.format(StatusCode.internalServerError, ErrorMessage.ERROR_REMOVING_USER_FROM_CONVO);
+    }
+  }
+}
+
+export default RemoveConversationMember;
